fix(api): preserve manifest overrides after reload
